fix(desktop): split organization members table
